+++ man-roxygen/param_store_benchmark_result.R
@@ -1,0 +1,2 @@
+#' @param store_benchmark_result (`logical(1)`)\cr
+#' Store benchmark result in archive?