+++ man-roxygen/param_store_models.R
@@ -1,1 +1,2 @@
 #' @param store_models (`logical(1)`).
+#' Store models in benchmark result?