+++ man-roxygen/param_store_resample_results.R
@@ -1,2 +1,0 @@
-#' @param store_resample_results (`logical(1)`)\cr
-#' Store resample results in archive?